+++ android_quickbrick_app/src/main/java/com/applicaster/ui/activities/MainActivity.kt
@@ -13,6 +13,7 @@ import androidx.annotation.RawRes
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.LifecycleObserver
 import androidx.lifecycle.OnLifecycleEvent
+import com.applicaster.analytics.AnalyticsAgentUtil
 import com.applicaster.plugin_manager.PluginManager
 import com.applicaster.plugin_manager.cmp.ConsentManager
 import com.applicaster.ui.R
@@ -20,8 +21,8 @@ import com.applicaster.ui.interfaces.HostActivityBase
 import com.applicaster.ui.interfaces.IUILayerManager
 import com.applicaster.ui.loaders.MainActivityPreloadSequence
 import com.applicaster.ui.quickbrick.QuickBrickManager
-import com.applicaster.ui.utils.CmpHookExecutor
-import com.applicaster.ui.utils.HookExecutor
+import com.applicaster.ui.utils.AppHookExecutor
+import com.applicaster.ui.utils.ConsentHookExecutor
 import com.applicaster.ui.utils.OrientationUtils.jsOrientationMapper
 import com.applicaster.ui.utils.OrientationUtils.nativeOrientationMapper
 import com.applicaster.ui.utils.OrientationUtils.normaliseOrientation
@@ -225,7 +226,7 @@ class MainActivity : HostActivityBase() {
         return when {
             hookPluginList == null || hookPluginList.isEmpty() -> Completable.complete()
             else -> Completable.create {
-                HookExecutor(this,
+                AppHookExecutor(this,
                         hookPluginList,
                         it,
                         isAppReady)
@@ -304,6 +305,7 @@ class MainActivity : HostActivityBase() {
         return createUIThreadCompletable { completableEmitter ->
             APLogger.debug(TAG, "UI ready...")
             initOrientationListener()
+            AnalyticsAgentUtil.logEvent(AnalyticsAgentUtil.APPLICATION_STARTED)
             setContentView(uiLayer!!.rootView) // simplistic approach, replace whole intro layout with RN layout
             completableEmitter.onComplete()
         }
@@ -321,7 +323,7 @@ class MainActivity : HostActivityBase() {
         return when {
             consentPlugins.isEmpty() -> Completable.complete()
             else -> Completable.create {
-                CmpHookExecutor(this,
+                ConsentHookExecutor(this,
                         consentPlugins,
                         it)
             }
