+++ android_quickbrick_app/src/main/java/com/applicaster/ui/loaders/PreloadStateManager.kt
@@ -11,7 +11,7 @@ class PreloadStateManager(private val defaultScheduler: Scheduler) {
         LOAD_DATA,
         VIDEO_INTRO,
         APPLICATION_READY_HOOK,
-        CMP_HOOK,
+        CONSENT_HOOK,
         UI_READY,
         RUNNING
     }
